official_models/ev3: refactoring
