+++ official_models/ev3/home/r3ptar/r3ptar.py
@@ -4,8 +4,8 @@
 Example LEGO® MINDSTORMS® EV3 R3ptar Program
 --------------------------------------------
 
-This program requires LEGO® EV3 MicroPython v2.0.
-Download: https://education.lego.com/en-us/support/mindstorms-ev3/python-for-ev3
+This program requires LEGO® EV3 MicroPython v2.0 downloadable at:
+https://education.lego.com/en-us/support/mindstorms-ev3/python-for-ev3
 
 Building instructions can be found at:
 https://www.lego.com/en-us/themes/mindstorms/buildarobot
@@ -28,83 +28,81 @@ class R3ptar:
 
     def __init__(
             self,
-            steer_motor_port: Port = Port.A,
-            drive_motor_port: Port = Port.B,
-            strike_motor_port: Port = Port.D,
+            steering_motor_port: Port = Port.A,
+            driving_motor_port: Port = Port.B,
+            striking_motor_port: Port = Port.D,
             touch_sensor_port: Port = Port.S1,
-            ir_sensor_port: Port = Port.S4, ir_beacon_channel: int = 1):
+            ir_sensor_port: Port = Port.S4,
+            ir_beacon_channel: int = 1):
         self.ev3_brick = EV3Brick()
 
-        self.steer_motor = Motor(port=steer_motor_port,
-                                 positive_direction=Direction.CLOCKWISE)
-        self.drive_motor = Motor(port=drive_motor_port,
-                                 positive_direction=Direction.CLOCKWISE)
-        self.strike_motor = Motor(port=strike_motor_port,
-                                  positive_direction=Direction.CLOCKWISE)
+        self.steering_motor = Motor(port=steering_motor_port,
+                                    positive_direction=Direction.CLOCKWISE)
+        self.driving_motor = Motor(port=driving_motor_port,
+                                   positive_direction=Direction.CLOCKWISE)
+        self.striking_motor = Motor(port=striking_motor_port,
+                                    positive_direction=Direction.CLOCKWISE)
 
         self.touch_sensor = TouchSensor(port=touch_sensor_port)
 
         self.ir_sensor = InfraredSensor(port=ir_sensor_port)
         self.ir_beacon_channel = ir_beacon_channel
 
-
-    def drive_by_ir_beacon(self, speed: float = 1000):
-        ir_beacons_pressed = set(self.ir_sensor.buttons(channel=self.ir_beacon_channel))
+    def drive_by_ir_beacon(
+            self,
+            speed: float = 1000,    # mm/s
+            ):
+        ir_beacons_pressed = \
+            set(self.ir_sensor.buttons(channel=self.ir_beacon_channel))
 
         if ir_beacons_pressed == {Button.LEFT_UP, Button.RIGHT_UP}:
-            self.drive_motor.run(speed=speed)
+            self.driving_motor.run(speed=speed)
 
         elif ir_beacons_pressed == {Button.LEFT_DOWN, Button.RIGHT_DOWN}:
-            self.drive_motor.run(speed=-speed)
+            self.driving_motor.run(speed=-speed)
 
         elif ir_beacons_pressed == {Button.LEFT_UP}:
-            self.steer_motor.run(speed=-500)
-
-            self.drive_motor.run(speed=speed)
+            self.steering_motor.run(speed=-500)
+            self.driving_motor.run(speed=speed)
 
         elif ir_beacons_pressed == {Button.RIGHT_UP}:
-            self.steer_motor.run(speed=500)
-
-            self.drive_motor.run(speed=speed)
+            self.steering_motor.run(speed=500)
+            self.driving_motor.run(speed=speed)
 
         elif ir_beacons_pressed == {Button.LEFT_DOWN}:
-            self.steer_motor.run(speed=-500)
-
-            self.drive_motor.run(speed=-speed)
+            self.steering_motor.run(speed=-500)
+            self.driving_motor.run(speed=-speed)
 
         elif ir_beacons_pressed == {Button.RIGHT_DOWN}:
-            self.steer_motor.run(speed=500)
-
-            self.drive_motor.run(speed=-speed)
+            self.steering_motor.run(speed=500)
+            self.driving_motor.run(speed=-speed)
 
         else:
-            self.steer_motor.hold()
+            self.steering_motor.hold()
+            self.driving_motor.stop()
 
-            self.drive_motor.stop()
-
-
-    def bite_by_ir_beacon(self, speed: float = 1000):
-        if Button.BEACON in self.ir_sensor.buttons(channel=self.ir_beacon_channel):
-            self.strike_motor.run_time(
+    def strike_by_ir_beacon(self, speed: float = 1000):
+        if Button.BEACON in \
+                self.ir_sensor.buttons(channel=self.ir_beacon_channel):
+            self.striking_motor.run_time(
                 speed=speed,
                 time=1000,
-                then=Stop.HOLD,
+                then=Stop.COAST,
                 wait=True)
 
-            self.strike_motor.run_time(
+            self.striking_motor.run_time(
                 speed=-speed,
                 time=1000,
                 then=Stop.COAST,
                 wait=True)
 
-            while Button.BEACON in self.ir_sensor.buttons(channel=self.ir_beacon_channel):
+            while Button.BEACON in \
+                    self.ir_sensor.buttons(channel=self.ir_beacon_channel):
                 pass
-
 
     def hiss_if_touched(self):
         if self.touch_sensor.pressed():
             self.ev3_brick.speaker.play_file(file=SoundFile.SNAKE_HISS)
-
 
     def main(self, speed: float = 1000):
         """
@@ -112,7 +110,7 @@ class R3ptar:
         """
         while True:
             self.drive_by_ir_beacon(speed=speed)
-            self.bite_by_ir_beacon(speed=speed)
+            self.strike_by_ir_beacon(speed=speed)
             self.hiss_if_touched()
             wait(1)
