+++ official_models/ev3/home/spik3r/.vscode/launch.json
@@ -8,7 +8,7 @@
 			"name": "Download and Run",
 			"type": "ev3devBrowser",
 			"request": "launch",
-			"program": "/home/robot/${workspaceRootFolderName}/main.py",
+			"program": "/home/robot/${workspaceRootFolderName}/spik3r.py",
 			"interactiveTerminal": false
 		}
 	]
